+++ examples/Blasphemos/check_size.py
@@ -1,0 +1,3 @@
+from PIL import Image
+img = Image.open(r"C:\Workspace\nesfab\nesfab\examples\Blasphemos\assets\silent_sorrow_9.png")
+print("Size:", img.size)